Enforce landscape orientation for scouting views
