+++ lib/screen-orientation.ts
@@ -1,8 +1,9 @@
+import { OrientationLock as ExpoOrientationLock } from 'expo-screen-orientation';
 import { NativeModulesProxy } from 'expo-modules-core';
 
 export const OrientationLock = {
-  PORTRAIT_UP: 'PORTRAIT_UP',
-  LANDSCAPE: 'LANDSCAPE',
+  PORTRAIT_UP: ExpoOrientationLock.PORTRAIT_UP,
+  LANDSCAPE: ExpoOrientationLock.LANDSCAPE,
 } as const;
 
 export type OrientationLockValue = (typeof OrientationLock)[keyof typeof OrientationLock];
